Final exam not done in class!!
Where was this work done?  Who did this code?  

After we spoke in the hallway, tonight, I thought you said you were going to get a drink of water and return.
When you did not return to the classroom to complete the final exam,
I therefore, assumed that you were withdrawing from the class, so I entered a grade of "W".

Since this work was not done in the classroom during the exam that I proctored,
I don't think I can count this code as work done for the in-class final exam for this course.

Furthermore, the final exam specified an array of  "Lobster" objects,
plus loops to manipulate each element of the array, but I do not see any arrays in this code (and there is no "report").
The lobster movement is not as specified (to the right and upward, with "bouncing"),
There is no animation (claws closing, as I illustrated in the classroom), and the "lobsters" do not interact with other objects.)

Therefore, your "Final-DAP1.java" code probably would not achieve a passing grade for the final exam.

Since you asked for more time to submit Project 9, I am still willing to accept that code tomorrow afternoon.
Your midterm grades were low (D-minus and C-minus), but a good Project 9 (including arrays, reports, sorting, etc.), 
might be sufficient to overcome the failing grade on the final exam, and earn a (barely) passing grade in the course.

So, please let me know whether or not you still want me to consider this "final-exam" code for grading.
If so, I will grade it then compute your letter-grade (ABCDF) and revise the records accordingly.
Otherwise, I will leave the "W" as your letter-grade for CST 112.

Please email me at "bam@suffolk.li" to let me know whether or not you want me to grade this code as your final exam for CST 112.